Expanded limits for some object vectors. Fixed numerous type casting warnings

+++ supercell-flash/source/flash/display_object/MovieClipFrame.cpp
@@ -12,7 +12,7 @@ namespace sc
 
 		void MovieClipFrame::save(SupercellSWF& swf) const
 		{
-			swf.stream.write_unsigned_short(elements_count);
+			swf.stream.write_unsigned_short((uint16_t)elements_count);
 			swf.stream.write_string(label);
 		}
